+++ actions/utils.py
@@ -1,0 +1,8 @@
+
+def show_more(name: str, url_prefix: str):
+    return {
+        "show_more": {
+            "name": name,
+            "url_prefix": url_prefix
+        }
+    }
